requirements.txt
pydantic>=2.5.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
requests>=2.31.0
urllib3>=2.1.0
beautifulsoup4>=4.12.2
lxml>=5.1.0
google-play-scraper>=1.2.4
# app-store-scraper pins requests==2.23.0 — install manually: pip install app-store-scraper --no-deps
# app-store-scraper>=0.3.5
praw>=7.7.1
PyGithub>=2.1.1
google-api-python-client>=2.111.0
playwright>=1.40.0
click>=8.1.7
rich>=13.7.0
tweepy>=4.14.0
pytest>=7.4.3
pytest-mock>=3.12.0
responses>=0.25.0
